+++ alimentare/src/Paginas/Alimentos.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Alimentos() {
+  return (
+    <div>Alimentos</div>
+  )
+}
