+++ alimentare/src/Paginas/Contacto.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Contacto() {
+  return (
+    <div>Contacto</div>
+  )
+}
