+++ alimentare/src/Paginas/Inicio.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Inicio() {
+  return (
+    <div>Inicio</div>
+  )
+}
