+++ alimentare/src/Paginas/Recetas.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Recetas() {
+  return (
+    <div>Recetas</div>
+  )
+}
